fix: extract _mock_pipeline fixture to shared helper (SonarCloud dup)
Move 15-line _mock_pipeline fixture from 5 test files into shared
conftest_public_api.install_mock_pipeline(). Removes ~155 duplicated
lines total (SonarCloud 3.49% → target ≤3%).
