+++ backend/tests/unit/conftest_public_api.py
@@ -1,13 +1,13 @@
 """Shared test helpers for public_api test files.
 
-Eliminates duplication of _make_result() and _mock_pipeline across
+Eliminates duplication of make_result() and mock_pipeline_agent across
 test_public_api_errors, test_public_api_facade, test_public_api_persistence,
 test_public_api_pipeline, test_public_api_session.
-
-Import with: from backend.tests.unit.conftest_public_api import make_result
 """
 
 from __future__ import annotations
+
+from collections.abc import Awaitable, Callable
 
 from backend.agents.executor_agent import PipelineResult
 from backend.agents.models import ExecutionPlan, PlanStep, ToolName
@@ -34,3 +34,34 @@ def make_result(
         "results": {"rows": [], "row_count": 0},
     }
     return result
+
+
+def make_fake_agent(
+    result_fn: Callable[..., PipelineResult] | None = None,
+) -> Callable[..., Awaitable[PipelineResult]]:
+    """Return a fake run_pilgrimage_agent coroutine for monkeypatching."""
+
+    async def _fake(
+        *,
+        text: str,
+        db: object,
+        model: object | None = None,
+        locale: str = "ja",
+        context: dict[str, object] | None = None,
+        on_step: object | None = None,
+    ) -> PipelineResult:
+        _ = (text, db, model, context, on_step)
+        if result_fn is not None:
+            return result_fn(locale=locale)
+        return make_result(locale=locale)
+
+    return _fake
+
+
+def install_mock_pipeline(monkeypatch: object) -> None:
+    """Monkeypatch run_pilgrimage_agent with a default fake."""
+    setattr_fn = monkeypatch.setattr
+    setattr_fn(
+        "backend.interfaces.public_api.run_pilgrimage_agent",
+        make_fake_agent(),
+    )
